refactor: watson-charater 위치 변경
