+++ src/components/watson/WatsonCharacter.tsx
