+++ src/features/survey/intro/Intro.tsx
@@ -5,9 +5,9 @@ import { AnimatePresence, m } from 'framer-motion';
 
 import CTAButton from '~/components/button/CTAButton';
 import StaggerWrapper from '~/components/stagger/StaggerWrapper';
+import WatsonCharacter from '~/components/watson/WatsonCharacter';
 import { defaultEasing } from '~/constants/motions';
 import { RESEARCHER_NAME } from '~/constants/name';
-import WatsonCharacter from '~/features/survey/intro/WatsonCharacter';
 import useBoolean from '~/hooks/common/useBoolean';
 import useStep from '~/hooks/step/useStep';
